Second Attempt
Complex Number class had a sign error in the subtraction method that has been fixed.  Seemingly not having transcript as an actual input for the Student Record class was bricking things.  The alternative is that I still need to define the initial objects for the Transcript components within the Student Record class despite them being dummy variables that get absorbed into the transcript list.